Crate symphonia_metadata

This crate implements a collection of metadata readers for standalone metadata formats, and read functions for metadata formats that are embedded into another container.

Standalone Metadata Formats

A standalone metadata format is one that exists independent of the media container.

This crate implements metadata readers (an object implementing the MetadataReader trait) for these metadata formats. A metadata reader may be registered with a probe for automatic detection. Each major standalone metadata format reader is implemented in a separate module.

Embedded Metadata Formats

An embedded metadata format is one that is embedded into the media container. This crate implements reading or parsing functions for these metadata formats in the embedded module.

Modules

ape

An APEv1 and APEv2 metadata reader.

embedded

Reading and parsing of metadata formats that are embedded into other containers.

id3v1

An ID3v1 metadata reader.

id3v2

An ID3v2 metadata reader.

utils

Common metadata utilities.