symphonia_metadata::id3v2

Module sub_fields

Key name constants for sub-fields of well-known ID3v2 frames.

For the exact meaning of these fields, and the format of their values, please consult the official ID3v2 specification.

Constants

AENC_OWNER

AENC_PREVIEW_LENGTH

AENC_PREVIEW_START

ATXT_EQUIVALENT_TEXT

ATXT_MIME_TYPE

COMM_LANGUAGE

COMM_SHORT_DESCRIPTION

COMR_CONTACT_URL

COMR_DESCRIPTION

COMR_MIME_TYPE

COMR_RECEIVED_AS

COMR_SELLER_LOGO

COMR_SELLER_NAME

COMR_VALID_UNTIL

CRM_DESRIPTION

CRM_OWNER

ENCRYPTION_METHOD_ID

ENCR_ENCRYPTION_DATA

ENCR_OWNER

GEOB_DESCRIPTION

GEOB_FILE_NAME

GEOB_MIME_TYPE

GRID_GROUP_DATA

GRID_OWNER

GROUP_ID

OWNE_PURCHASE_DATE

OWNE_SELLER_NAME

POPM_EMAIL

POPM_PLAY_COUNTER

POSS_POSITION_UNITS

PRIV_OWNER

SIGN_GROUP_ID

TXXX_DESCRIPTION

UFID_OWNER

USER_LANGUAGE

USLT_DESCRIPTION

USLT_LANGUAGE

WXXX_DESCRIPTION